requirements.txt
# Python dependencies for the contribution evaluation utility
requests>=2.0.0
jinja2>=3.0.0
PyYAML>=5.4.1
